# feat: optimize images to WebP format for better performance