--- a/components/customer-logos.tsx
+++ b/components/customer-logos.tsx
@@ -7,7 +7,7 @@ export function CustomerLogos() {
     <div className="flex flex-wrap justify-center items-center gap-8 md:gap-12">
       <div className="h-16 flex items-center justify-center">
         <Image
-          src="/images/logos/planeta.png"
+          src="/images/logos/planeta.webp"
           alt="Grupo Planeta"
           width={180}
           height={50}
@@ -16,7 +16,7 @@ export function CustomerLogos() {
       </div>
       <div className="h-16 flex items-center justify-center">
         <Image
-          src="/images/logos/alphaeditorial.png"
+          src="/images/logos/alphaeditorial.webp"
           alt="Alphaeditorial"
           width={180}
           height={50}
@@ -25,7 +25,7 @@ export function CustomerLogos() {
       </div>
       <div className="h-16 flex items-center justify-center">
         <Image
-          src="/images/logos/bajalibros.png"
+          src="/images/logos/bajalibros.webp"
           alt="BajaLibros.com"
           width={180}
           height={50}
@@ -34,7 +34,7 @@ export function CustomerLogos() {
       </div>
       <div className="h-16 flex items-center justify-center">
         <Image
-          src="/images/logos/latercera.png"
+          src="/images/logos/latercera.webp"
           alt="La Tercera"
           width={200}
           height={50}
@@ -43,7 +43,7 @@ export function CustomerLogos() {
       </div>
       <div className="h-16 flex items-center justify-center">
         <Image
-          src="/images/logos/marcombo.jpg"
+          src="/images/logos/marcombo.webp"
           alt="Marcombo"
           width={180}
           height={50}
@@ -52,7 +52,7 @@ export function CustomerLogos() {
       </div>
       <div className="h-16 flex items-center justify-center">
         <Image
-          src="/images/logos/penguin.png"
+          src="/images/logos/penguin.webp"
           alt="Penguin Random House Grupo Editorial"
           width={220}
           height={50}