… components/theme-options.tsx
… public/antarctica-digital-bookstore.webp
… public/ebooks-patagonia-interface.webp
… public/favicon.webp
… public/images/blue-theme.webp
… public/images/book-collection-1.webp
… public/images/book-collection-2.webp
… public/images/brand-customization.webp
… public/images/control-panel.webp
… public/images/hero-desktop.webp
… public/images/isologos/forbes-iso.webp
… public/images/isologos/penguin-iso.webp
… public/images/isologos/planeta-iso.webp
… public/images/isologos/santillana-iso.webp
… public/images/logos/alpha-cloud.webp
… public/images/logos/alphaeditorial.webp
… public/images/logos/amolca.webp
… public/images/logos/antartica-libros.webp
… public/images/logos/bajalibros.webp
… public/images/logos/bookshop.webp
… public/images/logos/canopus-lector.webp
… public/images/logos/carasycaretas.webp
… public/images/logos/christianbook.webp
… public/images/logos/ebooks-patagonia-logo.webp
… public/images/logos/forbes.webp
… public/images/logos/grupo-planeta.webp
… public/images/logos/hammurabi-logo.webp
… public/images/logos/la-tercera.webp
… public/images/logos/ladiaria.webp
… public/images/logos/latercera.webp