## Summary
- Convert 53 images from PNG/JPG/JPEG/GIF to WebP format
- Reduce total image size from 16MB to 3.2MB (80% reduction)
- Update all component references to use WebP images

## Performance Improvements
Based on Lighthouse testing in development mode:
- **Performance Score**: 27% → 99% ✅
- **Interactive Time**: 30.2s → 0.73s (97% reduction)
- **Total Blocking Time**: 2.6s → 92ms (96% reduction)
- **Largest Contentful Paint**: 588ms

## Changes Made
1. **Image Conversion**:
   - Team photos: 11.6MB → ~2MB
   - OG image: 2.7MB → 7.2KB
   - Marketplace interface: 570KB → 41KB
   - All logos converted with lossless compression to preserve transparency

2. **Code Updates**:
   - Updated all image references in TypeScript/React components
   - Fixed missing theme images by using existing assets
   - Added favicon.webp

3. **Documentation**:
   - Added comprehensive `IMAGE_OPTIMIZATION_PLAN.md` with conversion details
   - Documented actual results and next steps

## Test Plan
- [x] All images load correctly in development
- [x] Lighthouse performance test shows significant improvement
- [ ] Verify images load correctly in production build
- [ ] Test on different browsers (WebP is supported by 96%+ of browsers)
- [ ] Verify OG image displays correctly in social media previews

## Next Steps
After merging and verifying in production:
1. Remove original PNG/JPG files to save additional space
2. Delete `public/images-backup` directory
3. Consider implementing responsive images with srcset for further optimization

🤖 Generated with [Claude Code](https://claude.ai/code)